fixed warnings
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@4944 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcaqgui2dfactory/pathpainter.cpp
@@ -91,7 +91,7 @@ void PathPainter::paint( QPainter *painter, int z )
         futureWpColor = hydroqguielementutil::getTransparentVersion(baseColor,0.75);
     }
     
-    for ( unsigned int i=0; i < guiPath_.size(); i++)
+    for ( int i=0; i < (int)guiPath_.size(); i++)
     {
         QColor fillColor;
         QColor drawColor;
